topic_names -> topics

+++ kafkarestv3/model_list_mirror_topics_response_data.go
@@ -18,5 +18,5 @@ type ListMirrorTopicsResponseData struct {
 	NumPartitions int32 `json:"num_partitions"`
 	MirrorLags []MirrorLag `json:"mirror_lags"`
 	MirrorStatus MirrorTopicStatus `json:"mirror_status"`
-	StateTimeMs int32 `json:"state_time_ms"`
+	StateTimeMs int64 `json:"state_time_ms"`
 }
